+++ metadata.rb
@@ -1,7 +1,7 @@
-name 'pgpool-cookbook'
+name 'pgpool'
 maintainer 'Jens Jorritsma'
 maintainer_email 'jens.jorritsma@gamil.com'
 license 'All rights reserved'
-description 'Installs/Configures pgpool-cookbook'
+description 'Installs/Configures pgpool'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version '0.1.0'