+++ test/cookbooks/pgpool-test/metadata.rb
@@ -5,4 +5,4 @@ license 'apachev2'
 description 'Tests pgpool Resources'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version '0.1.0'
-depends 'pgpool-cookbook'
+depends 'pgpool'