+++ CHANGELOG.md
@@ -1,6 +1,6 @@
-pgpool-cookbook CHANGELOG
+pgpool Cookbook CHANGELOG
 =========================
 
 0.1.0
 -----
-- Jens Jorritsma - Initial release of pgpool-cookbook
+- Jens Jorritsma - Initial release of pgpool